refactor(openrouter): carry routing as opaque driver option
Move OpenRouter API types (routing config, presets, request/response
extensions) from everruns-provider into the openrouter driver crate.
Core and provider now pass routing through LlmCallConfig.driver_options
under the openrouter/routing key; the driver deserializes it at request
time and drops missing/empty routing, preserving wire behavior.
