+++ crates/drivers/anthropic/tests/parallel_tool_calls_live.rs
@@ -57,7 +57,7 @@ fn config_with(parallel: Option<bool>) -> LlmCallConfig {
         provider_opaque_context: None,
         tool_search: None,
         prompt_cache: None,
-        openrouter_routing: None,
+        driver_options: Default::default(),
         parallel_tool_calls: parallel,
         volatile_suffix_len: 0,
         extra_headers: Vec::new(),
