+++ crates/drivers/fireworks/tests/chat_live.rs
@@ -41,7 +41,7 @@ async fn fireworks_chat_streams_response() {
         provider_opaque_context: None,
         tool_search: None,
         prompt_cache: None,
-        openrouter_routing: None,
+        driver_options: Default::default(),
         parallel_tool_calls: None,
         volatile_suffix_len: 0,
         extra_headers: Vec::new(),
