+++ crates/drivers/meta/src/driver.rs
@@ -193,7 +193,7 @@ mod tests {
             provider_opaque_context: None,
             tool_search: None,
             prompt_cache: None,
-            openrouter_routing: None,
+            driver_options: Default::default(),
             parallel_tool_calls: Some(false),
             volatile_suffix_len: 0,
             extra_headers: vec![],
