+++ crates/drivers/openai/tests/astra_compaction_live.rs
@@ -35,7 +35,7 @@ async fn astra_compaction_preserves_facts_constraints_and_continuation() {
         provider_opaque_context: None,
         tool_search: None,
         prompt_cache: None,
-        openrouter_routing: None,
+        driver_options: Default::default(),
         parallel_tool_calls: None,
         volatile_suffix_len: 0,
         extra_headers: vec![],
