+++ crates/engine/src/execution/reason/tests.rs
@@ -650,7 +650,7 @@ fn test_build_request_options_for_openai_prompt_cache() {
             strategy: PromptCacheStrategy::Auto,
             gemini_cached_content: None,
         }),
-        openrouter_routing: None,
+        driver_options: Default::default(),
         parallel_tool_calls: None,
         volatile_suffix_len: 0,
         extra_headers: Vec::new(),
@@ -690,7 +690,7 @@ fn test_build_request_options_for_gemini_explicit_cache() {
             strategy: PromptCacheStrategy::Auto,
             gemini_cached_content: Some("cachedContents/demo-cache".to_string()),
         }),
-        openrouter_routing: None,
+        driver_options: Default::default(),
         parallel_tool_calls: None,
         volatile_suffix_len: 0,
         extra_headers: Vec::new(),
@@ -730,7 +730,7 @@ fn test_build_request_options_omits_gemini_cache_flag_when_disabled() {
             strategy: PromptCacheStrategy::Auto,
             gemini_cached_content: Some("cachedContents/demo-cache".to_string()),
         }),
-        openrouter_routing: None,
+        driver_options: Default::default(),
         parallel_tool_calls: None,
         volatile_suffix_len: 0,
         extra_headers: Vec::new(),
