+++ crates/provider/tests/openresponses_protocol_wire.rs
@@ -37,7 +37,7 @@ fn config(model: &str) -> LlmCallConfig {
         provider_opaque_context: None,
         tool_search: None,
         prompt_cache: None,
-        openrouter_routing: None,
+        driver_options: Default::default(),
         parallel_tool_calls: None,
         volatile_suffix_len: 0,
         extra_headers: Vec::new(),
