+++ crates/server/src/domains/observers/judge.rs
@@ -235,7 +235,7 @@ impl JudgeClient for LlmJudgeClient {
             provider_opaque_context: None,
             tool_search: None,
             prompt_cache: None,
-            openrouter_routing: None,
+            driver_options: Default::default(),
             parallel_tool_calls: None,
             volatile_suffix_len: 0,
             extra_headers: Vec::new(),
