+++ crates/test-support/tests/reason_atom_test.rs
@@ -2869,7 +2869,7 @@ async fn test_driver_registry_integration() {
         provider_opaque_context: None,
         tool_search: None,
         prompt_cache: None,
-        openrouter_routing: None,
+        driver_options: Default::default(),
         parallel_tool_calls: None,
         volatile_suffix_len: 0,
         extra_headers: Vec::new(),
