+++ integrations/openrouter-workspace/Cargo.toml
@@ -19,6 +19,7 @@ default = []
 [dependencies]
 everruns-core.workspace = true
 everruns-provider.workspace = true
+everruns-openrouter = { path = "../../crates/drivers/openrouter" }
 async-trait.workspace = true
 reqwest.workspace = true
 serde.workspace = true
